Add: NASL builin function open_sock_kdc
