+++ rust/nasl-builtin-network/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "nasl-builtin-network"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+nasl-builtin-utils = { path = "../nasl-builtin-utils" }
+nasl-syntax = { path = "../nasl-syntax" }
+storage = { path = "../storage" }
+
+[dev-dependencies]
+nasl-interpreter = { path = "../nasl-interpreter" }
