+++ rust/nasl-builtin-network/src/lib.rs
@@ -1,0 +1,5 @@
+// SPDX-FileCopyrightText: 2024 Greenbone AG
+//
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+pub mod socket;
