+++ rust/nasl-builtin-utils/src/context.rs
@@ -288,7 +288,11 @@ impl Default for Register {
         Self::new()
     }
 }
-use std::collections::HashMap;
+use std::{
+    collections::HashMap,
+    net::{TcpStream, UdpSocket},
+    sync::{Arc, Mutex, RwLock},
+};
 type Named = HashMap<String, ContextType>;
 
 /// NaslContext is a struct to contain variables and if root declared functions
@@ -328,6 +332,12 @@ impl NaslContext {
     }
 }
 
+pub enum Connection {
+    Closed,
+    TCP(TcpStream),
+    UDP(UdpSocket),
+}
+
 /// Configurations
 ///
 /// This struct includes all objects that a nasl function requires.
@@ -364,6 +374,7 @@ impl<'a> Context<'a> {
             retriever,
             loader,
             executor,
+            connections: vec![],
         }
     }
 
@@ -419,6 +430,18 @@ impl<'a> Context<'a> {
                 None => NaslValue::Null,
             })
             .map_err(|e| e.into())
+    }
+
+    /// Get a Connection
+    pub fn get_connection(&self, id: usize) -> Option<&Connection> {
+        self.connections.clone().read().unwrap().get(id);
+    }
+
+    /// Save a new connection in the context
+    pub fn new_connection(&self, con: Connection) -> usize {
+        let ret = self.connections.len();
+        self.connections.push(con);
+        ret
     }
 }
